+++ src/main/java/frc/robot/commands/auto_commands/PositionOne.java
@@ -2,6 +2,8 @@ package frc.robot.commands.auto_commands;
 
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import frc.robot.commands.*;
+import frc.robot.commands.hopper.HopperMotorRun;
+import frc.robot.commands.shooter.ShooterRun;
 
 public class PositionOne extends SequentialCommandGroup {
