+++ src/main/java/frc/robot/commands/control_panel/ControlPanelExtend.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.commands.control_panel;
 
 import frc.robot.Robot;
 import edu.wpi.first.wpilibj2.command.InstantCommand;
