+++ src/main/java/frc/robot/commands/control_panel/ControlPanelMotorStop.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.commands.control_panel;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.Robot;
