+++ src/main/java/frc/robot/commands/hopper/HopperMotorStop.java
@@ -1,17 +1,17 @@
-package frc.robot.commands;
+package frc.robot.commands.hopper;
 
 import frc.robot.Robot;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 
-public class HopperMotorRun extends CommandBase {
+public class HopperMotorStop extends CommandBase {
 
-    public HopperMotorRun() {
+    public HopperMotorStop() {
         addRequirements(Robot.hopperSubsystem);
     }
 
     @Override
     public void execute() {
-        Robot.hopperSubsystem.run();
+        Robot.hopperSubsystem.stop();
     }
 
     @Override
