+++ src/main/java/frc/robot/commands/hopper/HopperSequential.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.commands.hopper;
 
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
