+++ src/main/java/frc/robot/commands/shooter/ShooterRun.java
@@ -5,7 +5,7 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands;
+package frc.robot.commands.shooter;
 
 import frc.robot.Robot;
